+++ ts/src/program/namespace/rpc.ts
@@ -21,7 +21,11 @@ export default class RpcFactory {
       const tx = txFn(...args);
       const [, ctx] = splitArgsAndCtx(idlIx, [...args]);
       try {
-        return await provider.send(tx, ctx.signers ?? [], ctx.options);
+        return await provider.sendAndConfirm(
+          tx,
+          ctx.signers ?? [],
+          ctx.options
+        );
       } catch (err) {
         throw translateError(err, idlErrors);
       }